Moved ItemChanges to its own file
